// frontend/src/app/api/model/planRequest.ts
/**
 * QuickFlow API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PlanItemSource } from './planItemSource';


export interface PlanRequest { 
    title: string;
    estimatedDurationMinutes: number;
    startDateTime: string;
    /**
     * Must be after startDateTime
     */
    endDateTime: string;
    priorityOrder: number;
    items: Array<PlanItemSource>;
}